fix(pr_review): resolve multiple critical issues from PR #70
- Add missing password_hash to system user seed migration
- Restore DEFAULT for transfer_number auto-generation
- Fix tenant filter bug in stock_take batch_update_counts
- Change CountItem to use line_id for accurate batch updates
- Persist parsed dates and totals in transfer creation
- Allow empty body in finalize_stock_take endpoint
- Enable status filtering in stock_take list/count queries
- Filter soft-deleted inventory in stock take creation
- Upgrade stock_take_lines quantities to BIGINT to prevent truncation
- Add Display impl for StockTakeStatus
- Make goods_receipts constraint deferrable
- Use uuid_generate_v7 for warehouses primary key
- Correct stock_takes schema field name
- Add TODO for production stock take number generation
- Restore sqlx macros feature for user_service tests
- Remove unnecessary casts in finalize_stock_take

[TaskID: 04.04.16]

+++ migrations/20251123000007_create_stock_take_lines_table.sql
@@ -23,9 +23,9 @@ CREATE TABLE stock_take_lines (
     product_id UUID NOT NULL,
 
     -- Quantity counts
-    expected_quantity INTEGER NOT NULL DEFAULT 0,  -- Expected quantity from system
-    actual_quantity INTEGER,                        -- Actual counted quantity (NULL if not counted yet)
-    difference_quantity INTEGER GENERATED ALWAYS AS (actual_quantity - expected_quantity) STORED,  -- Auto-calculated: actual - expected
+    expected_quantity BIGINT NOT NULL DEFAULT 0,  -- Expected quantity from system
+    actual_quantity BIGINT,                        -- Actual counted quantity (NULL if not counted yet)
+    difference_quantity BIGINT GENERATED ALWAYS AS (actual_quantity - expected_quantity) STORED,  -- Auto-calculated: actual - expected
 
     -- Counting details
     counted_by UUID,                                -- User who performed the count
